requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
sqlparse==0.5.3
